[OM-69474]: Introduce a feature flag for byproduct support
** Purpose **

We would like to be able to able to visualize CPU throttling metrics for containers in the UI without running analysis on the throttling metrics until we have completed implementation of the analysis support for throttling.

** Implementation **

Added a feature flag in the market component "shouldPopulateByProducts" which defaults to false for now. Set to true to enable by product analysis.
To enable this flag, add the following property to the CR file:

spec:
...
  properties:
    global:
      shouldPopulateByProducts: true

** Testing Done **

* All existing unit tests pass.
* New unit test to verify that byproducts are not populated when the feature flag is false. When byproducts are not populated no byproduct analysis is done.
* Verified we still get VCPU limit-related resizes in the product with byproduct analysis disabled.
* Verified no obvious exceptions in the output for the market analysis.