[OM-68158] Moved the `LicenseXmlDTO` class outside `LicenseDeserializer`.
There were 4 classes in `LicenseDeserializer.java` and ~500 lines. Since
`LicenseXmlDTO` is largely independent, it was moved to it's own file to
reduce the number of max classes/file to 2 and max file size to ~300
lines.

`LicenseXmlDTO` was public but never used outside the
`com.vmturbo.licensing.utils` package so it was also given package
access.