[OM-68158] Replaced log4j2 test configuration to properly suppress logs.
Logging shouldn't be enabled by default in unit tests as these are run
in an unattended way. Moreover the errors are expected and confusing to
see. Instead the logs generated by Maven and JUnit themselves should
remain uncluttered.

This file borrowed from `com.vmturbo.commons` is intended to allow one
to switch on logging easily when needed.