[OM-68158] Enabled Checkstyle for LicenseDeserializer and fixed violations.
- Added missing Javadoc comments.
- Added missing Javadoc tags.
- Fixed comment formatting.
- Suppressed generation of an unused default constructor.
- Removed whitespace after casts.
- Removed redundant `public` modifier.

Also:

- Fixed a grammar mistake.
- Rewrapped a few lines.

Comments may require the attention of an expert on this part of the
codebase to provide more info that would actually make them useful. I
could only bring them up to the requirements of Checkstyle.

`LicenseXmlDTO` used to be in `LicenseDeserializer` and Checkstyle
violations in it were fixed as well.