[OM-68158] Added unit tests for LicenseXmlDTO.
A parametrised unit test was added with 3 different inputs: two for the
existing license files and one for a new license that contains the
customer ID. Reflection is used to ensure all license fields are tested.

The test provides 96% method coverage and 91% line coverage to
`LicenseXmlDTO`. The only missed method/lines is `setFeatures` which is
only called when deserialising license files that use `features`
elements instead of `feature` elements. I'm not sure if this intended to
be supported or the method should be removed.

Also modified `pom.xml` to include the library for parametrised tests
and to do some cleanup.