[OM-70200]: Ensure correct population of commodity index
**Problem**

Testing showed that that Container plan node capacity values differed greatly before and after plans. Investigation revealed that the cause was that the CommodityIndex used to convert commodity values from market Traders to projected TopologyEntities was empty when it should not have been.

**Solution**

The reason the `CommodityIndex` was empty is because the getter for `CommodityIndex` actually triggers a memoized call to create the `CommodityIndex`. This allows the creation of the `CommodityIndex` to be deferred until it is actually needed. A recent change to add a feature flag moved up the first call to get the `CommodityIndex` much earlier than it used to be called, triggering the `CommodityIndex` to be created before the data structures it relies on are populated with any real data. As a result, the `CommodityIndex` was empty.

The fix is to defer the call to get the `CommodityIndex` again until the data structure it relies on has the necessary data. Since the ordering dependency on the creation of `CommodityIndex` is unclear and it would be easy for someone to make a similar mistake again in the future, I have also added in a check that will throw an `IllegalStateException` if someone again moves the creation of `CommodityIndex` too early in the lifecycle of the `TopologyConverter`.

**Testing Done**
* Added a new unit test to ensure we throw an `IllegalStateException` when creating commodity index when the state of the `TopologyConverter` is insufficient to permit it.
* Verified that the bug was reproducible before my fix and not reproducible after my fix.
* Updated existing unit tests that failed after these changes to account for the new constructor signature to `TopologyConverter`.