[OM-69909]: Resolve Reactor library version conflict causing errors in Azure discovery when running Voltron.
Reactor library is a direct dependency for XL and an indirect dependency for Azure probe. XL uses 3.0.4.RELEASE version while Azure probe transitively uses 3.3.10.RELEASE. When Voltron starts Azure discovery it loads 3.0.4 version first and fails later on because Azure assumes 3.3.10 version.
The fix is to advance reactor version in XL to 3.3.10.RELEASE.