[OM-70406]: Fix swagger unpack to work in non-SNAPSHOT branches
**Problem**

One step in the swagger unpacking build step deletes a temporary directory. The code that identified the temporary files to delete was looking for the presence of "SNAPSHOT" in the directory path. This worked fine in SNAPSHOT-based branches, but caused release branches where the SNAPSHOT tag is removed to fail.

**Solution**

Instead of looking for "SNAPSHOT", look for the com.vmturbo.api.version in the path which is probably a better thing to look for anyway.

**Testing Done**

Reproduced build failure in local version of 8.1.6 branch. Tested that fix resolves the issue on my local in 8.1.6. Also tested that building staging works with this fix too.