[OM-68650] Differentiate between zero counts and cases where we don't have the data in counting stats
In the case where we haven't expanded the scope to include a particular enitity, we will no longer try to count that entity type and return a (possibly erroneous) zero result.
However, we still provide zero counts in cases where we do try to expand the scope to a particular type and still find none in the scope. This is made possible by providing additional information (derivedEntityTypes) in the request to History component.