[OM-59021]: Migrate the extneral-api-stubs project from the OpsManager repo to the Api project repo
When we recently split the API definition to a separate repository, the external-api-stubs module did not come with it. This leaves it using outdated swagger to generate its test classes. If XL is to continue to depend on this project, it either needs to move to the API project or XL. After discussion with key stakeholders, we decided to move it to the API project.

This commit is to update XL to depend on the newly-migrated module.

There are corresponding changes in API to  migrate the module, as well as changes in OpsManager to remove the dead code.