[OM-70857]: Add safeguards in case t8c.sh is run more than once
Problem: If t8c.sh is re-run while using local storage, the system is left in an inconsistent state.
Specifically, we erase the kubernetes data (including the master key secret) but not the pod data (some of which was encrypted using the master key). This leaves the system in a state where it can not decrypt the existing secrets. The reason for this is that we did not umount the volume housing the pod data before trying to wipe it.

Solution:

Fix the script so that it properly deletes the pod data as well as the kubernetes data when re-running t8c.sh. In this case, there will be a new master key and no pod data. The new master key will be used to encrypt new sensitive pod data and will not prevent the system from coming up cleanly.