[OM-70686]: Populate name in the workflow protobuf message sent for creating/updating workflows
Problem
=====
Only one user created workflow can be created. The DataIntegrityViolationException exception was thrown when adding more than one user-created workflow.

Diagnosis
=========
Workflow identity provider uses the target id and name to assign an id to a workflow. For user-created workflows, we do not populate the name in the workflow object. Therefore, all of them get the same id.

Solution
========
When creating user-created workflows, use a random uuid for name.