[OM-70659]: Optimize CpuScalingFactorPostStitchingOperation
**Problem**

The `CpuScalingFactorPostStitchingOperation` takes a very long time to complete at Citi.

**Investigation**

The `CpuScalingFactorPostStitchingOperation` sets a scaling factor to normalize CPU performance for workloads running on physical hardware where raw MHz is not a good measure of CPU performance (for example, an old 3GHz processor and a new 3GHz processor may both be 3GHz but you will get much better performance on the new one because of different processor architectures, etc.)

The operation needs to propagate the scaling factors set on on-prem physical hosts recursively to all consumers of CPU-related commodities. This includes container-related entities with commodities like CPU limits, requests, and quotas. Further, entities like namespaces and workload controllers connected to multiple physical hosts should get the maximum of all the connected host scaling factors to prevent the market computing an updated usage number less than zero (triggering an `IllegalArgumentException`).

The current implementation propagates updates on workload controller and namespace entities (which require the max) to both providers and consumers. We protect against infinite recursion by only updating the consumer direction if it is the first time we update an entity. Unfortunately, for highly connected namespaces, this results in updating the connected namespace and workload controller many times. On the citi topology, we called the update function over 12 million times. Even though each update is fast, calling it so many times resulted in a big performance hit.

**Solution**

The prior implementation for the operation ensured that workload controller and namespace scaling factors matched by propagating the update in both the provider and consumer directions on namespace and workload controller. However, after talking with @YueWu we agreed that it should not be necessary to update in the consumer direction on these entities, elimiinating many needless updates.

In terms of behavioral impact, this will mean that the namespace will get the maximum scaling factor of all hosts connected to any of its connected pods. WorkloadControllers, instead of getting the same value as their connected namespace, will now get the scaling factor of all hosts connected to any pods they are connected to. This should not trigger any `IllegalArgumentException`s during market analysis. If it does, we can add another simple post-stitching operation to run after the `CpuScalingFactorPostStitchingOperation` which just sets all WorkloadController scaling factors to match their connected Namespace. Additionally, the market team has added in defensive logic in the market to prevent `IllegalArgumentException`s from stopping market analysis anymore. Instead, it will just prevent action generation on a single Container scaling group.

This change reduced the number of times that `updateScalingFactorForEntity` from over 12 million down to 670 thousand.

**Testing Done**
1. Updated unit tests to ensure we do not perform unnecessary updates on container entities.
2. Verified that `CpuScalingFactorPostStitchingOperation` runs faster on the Citi topology.
  1. Performance on laptop: Took ~3.5 minutes before optimization to ~3.5 seconds after optimization
  2. Performance on an appliance in the lab: Took 10+ minutes before optimization, Took ~30s after optimization
3. Verified that we get no `IllegalArgumentException`s in the market during analysis after this update on the Citi topology as well as a smaller in-house container topology.