[OM-72018]: Use 'Retain' reclaim policy for persistent volumes using local storage
Background
==========
Persisent Volumes (PVs) in Kubernetes use a reclaim policy to tell the cluster what to do with the PV when the PVC (PV claim) is no longer active. In our Gluster (legacy) configuration, we used the "Delete" policy, which is the default and made sense because we were using a dyanmic provisioner (Heketi).

Problem
=======
In our newer configuration using local storage, we do not have a dynamic provisioner. Therefore the 'Delete' reclaim policy is not supported and causes an error when the PVC is removed.

Solution
=======
Change the reclaim policy in our local storage configuration to 'Retain'. This is the only officially supported setting for local storage configurations.

The effect of this is that, when the PVC is released (meaning our pods no longer need the PV), the PV will remains along with all data on the filesystem of the VM. However, the PV can not be reused by other pods without administrator intervention. It is not an expected situation to try to re-use a persistent volume--at least in production. In a test environment, this could happen when flipping feature flags back and forth.