[OM-72213]: Enable Flag for Historical Pending Action Counts by default
Enables the ability to count the number of Historical Pending Actions so that I can use the counts to create a report to show the number of historical pending actions over time.

We have previously developed this feature in a previous story, but it is currently disabled by default behind a feature flag. This story is to complete testing around the new feature and then enable it by default.


(cherry picked from commit 28ea67b7cad6f5e147e9c29bb9728fdaeb19bcb0)