[OM-72300]: Enable connection pooling for MariaDB (MySQL) connections
Problem
=======
We had started seeing gRPC related errors observed across all components due to:

     java.sql.SQLNonTransientConnectionException: Could not connect to address=(host=db)(port=3306)(type=master) : Cannot assign requested address (connect failed)

This seems to happen when we have many (tens of thousands) of database queries happen in a short period of time, which apparently exhausts the number of available sockets in the underlying operation system. Note that while most database queries finish very quickly, it takes minutes for the socket to become available for reuse.

Solution
========
Enable connection pooling in the MariaDB adapter library, so that we reuse connections for multiple requests rather than always making a new connection.