[OM-73183]: Fix HeapDump REST controller
**Background**

When working on the security fix for bug 72452 I moved the HeapDump functionality into a new gRPC service so that it could be separately created/not created based on the `removeHeapDumpService` flag. This resulted in a new `HeapDumpServiceController` that required a separate Spring configuration bean to be instantiated. I did all my testing via gRPC so I missed the fact that the REST controller for heap dumps was no longer available. The REST controller is important in being able to trigger heap dumps within the container (which will be done in production or appliances) because `grpc_cli` is not available there.

**Implementation**

- Create a `HeapDumpServiceController` Spring bean in `BaseVmtComponentConfig`.
- I slightly refactored the intiialization process for the `HeapDumpRpcService` and got rid of the no longer necessary factory.

**Testing Done**

- Ensured that we can trigger heap dump in a component via a REST call when `removeHeapDumpService` flag is false
- Ensured that we cannot trigger heap dump in a component via a REST call when `removeHeapDumpService` flag is true
- Ensured that gRPC requests to the HeapDumpRpcService still work.