[OM-72343]: Historical throttling statistics
<h1>Intent</h1>

During analysis, throttling is treated as a by-product of the VCPU commodity to determine when to resize up and down VCPU.

We want to retain metrics on historical throttling in a way that permits fast resizing up when a spike in throttling occurs while also having a sufficiently long memory for any throttling spikes in the past. To achieve this, we maintain fast and slow moving statistics and take the max between the two of them to send to market for analysis.

When the VCPU commodity capacity changes, we cannot retroactively update historical data collected at other capacities to reflect the new capacity. Instead, we maintain separate moving statistics for each capacity.

Based on experiments in a testbed, we determined that throttling starts to have a noticable impact on application response time when it reaches a utilization of about 5% or more. As a result, we set a utilThreshold of 5% for the throttling commodity (not in this review), leading to a desired state of 3.5% (0.7 desired state * 5% = 3.5%). To prevent resizing back down VCPU to a level that introduces throttling after resizing up to eliminate it, we set a minThreshold on the VCPU commodity by linearly interpolating data points around the 3.5% target throttling utilization (the market team asked for this approach instead of sending multiple data points for analysis of a single commodity).

<h1>Implementation</h1>

We leverage the `HistoryAggregator` abstractions for maintaining history for the throttling and VCPU commodities.

<h1>Testing Done</h1>

Testing in progress. Added some basic unit tests for `VcpuThrottlingSampler`.