[OM-73952]: Introduce a flag to control whether database connection pooling is used and set it to false by default for 8.3.0
Introduce a flag to control whether database connection pooling is used. This can be set per-component or globally.

This will default to false for 8.3.0, to give us more time to assess the performance impact of DB connection pooling, and identify possible connection leaks in our existing code.

This change does not affect the History component.