[OM-74168]: Fix log statement for MovingStatistics commodities edited
**Intent**

There is a log statement from the `HistoryAggregator` class when it completes its work in the `HistoryAggregationStage` which reports the number of commodities edited by each editor that it runs. The `MovingStatisticsEditor` was incorrectly reporting 0 commodities edited even when it was editing many commodities.

Also fix an issue with the `Timers` utilities where they were not thread safe. This was problematic when I was attempting to get performance metrics from the concurrent code in the `HistoryAggregator`-related classes.

**Implementation**

The `HistoryAggregator` log was generated by iterating messages by their associations with each editor's class name. The `MovingStatisticsEditor` was associating commodity edits with a different string than its class name. As a result, even though it was recording its commodity edits, they weren't being printed. The fix is to update the code to associate commodity edits with the `MovingStatisticsEditor` class name.

The fix for the timers is to use a `ConcurrentHashMap` instead of a regular `HashMap` and `AtomicLong`s instead of regular primitive longs.

**Testing Done**

Old log message:
```
topology-processor-64c6b8cb67-4n 2021-08-17 11:07:29,950  INFO [HistoryAggregator]      : History aggregator commodities modified: TimeSlotEditor:0 MovingStatisticsEditor:0 PercentileEditor:1131 in 761.1 ms
```

New log message:
```
2021-08-17 15:05:39,098  INFO [HistoryAggregator]	: History aggregator commodities modified: MovingStatisticsEditor:10530 TimeSlotEditor:0 PercentileEditor:46964 in 5.763 s
```

Timers were tested by experiencing a ConcurrentModificationException while profiling the parallel `HistoryAggregator` code. After making the Timers threadsafe, these exceptions are gone.