[OM-74660]: Enable gRPC Server Reflection
To provide developers the ability to turn on gRPC server reflection as and when needed.

Robot tests require communication with the gRPC server to test the functionality of the different XL components. However, the robot client needed to be aware of the `protobuf` files to send any calls to the gRPC server.
This change introduces an optional attribute to enable gRPC Server Reflection, which allows clients to list the calls offered by the server, and make any calls, without knowing the internal details of the gRPC calls in advance. This also negates the necessity for the robot client to copy all protobuf dependencies for each version of XL, or anytime a change occurs.

By default, the property is disabled to make sure there are no performance impacts.

Testing with Docker-compose:

```
defaultProperties:
    global:
        enableGrpcReflection: true
```
And run the group component like so:
```
$ cd com.vmturbo.group.component
$ docker-compose up
```
Or to run this in XL, make a change to the crd file:
```
   properties:
     global:
       enableGrpcReflection: true
```