[OM-75715]: Do the filtering of targets using UI category rather than the backend category
Fixed the filtering by category in targets service

The targets have two categories: categories and uicategories. We should use UI category for all the operations we do in the API level as the customer is not aware of the other category. In the targets service filtering we were comparing based on cateogry rather uicategory that got fixed in the change.