[OM-76350]: Disable a test that fails occasionally in the build server
Disable a test that fails once in a while in the build server until we figure out why it is happening.

We have this test that fails once in a while in the build server. I have tried to reproduce it locally but I had no success. I also looked into the code and cannot understand why it is failing but I had no success. I made this change to disable the test until we can figure out why the test is failing.