[OM-77180]: bump the version of outdated maven-jar-plugin to 3.2.0
Fix the issue that outdated version of maven-jar-plugin is incompatible with newer version of maven

I had an issue that the containers that I build exited with class not found exception. With Joseph's help, I figured out the issue is that we are using a very old maven-jar-plugin that is not compatible with newer mavens. So newer mavens do not add a proper classpath to the jar manifest file. See the corresponding issue raised in apache Jira here https://issues.apache.org/jira/browse/MJAR-279?page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel&focusedCommentId=17434264#comment-17434264
I think we should soon bump up the maven-jar-plug from 2.5 which is for 2012 to a more recent version.

One additional fixes:
1. There was place in the code that expect Specification-Vendor in the manifest file to have three version parts. However, it will always have two parts (major and minor) https://maven.apache.org/shared/maven-archiver/ . I updated the code two handle bother three and two parts.