[OM-77963]: Fix test CachingMemberCalculatorTest#testConcurrentUpdateOnRegroup that fails intermittently on the build machine
Fix a flaking a test that has a race condition

This is the second attempt the fix this test. First I will give the information about the first time fix and then I will information about the second time fix.

This test that is failing was added to make sure that if we have a change in the groups in the cache during a regroup, we don't lose those changes and we re-apply those changes after the regroup operation is finished. This test is failing because of a race condition. The race condition can be explained by explaining how does the cache works. When a regroup is started we turn on a flag that enables recording the changes during the regroup and then we go to database to get the group definition to get their members. The test is implemented as follows: the logic for getting the group members are overriden so we can lock it for a period of time. During that period, we make changes to the cache. After that period, we make release the lock and then we make sure the cache has the correct value. The race condition happens in the test when we perform changes before enabling that flag. In reality, the db will have the changes and we don't need to record those changes. However, that will not happen in our test because we are mocking the db interaction part and always returning fixed results. Therefore, we lose that change and that results in the test failing.

For the initial fix, I added a  count down latch. The count down latch will decrement when we are in getting new memebers logic. The testing will wait until the count down latach becomes zero or a timeout period passed. This way, we can be sure that we apply  the changes the regroup flag is true and we are capturing the changes.

The issue with the initial fix was that we are calling the regroup twice and the test should wait for second regroup to happen. Unfortunately, I did not notice  that when I did the first fix and the issue and the first regroup release latch that was used for waiting. This made my changes ineffective for synchronsing the code for second regroup. This can fixed by initializing the count down latch by 2. The first regroup will change it to 1 and therefore, the code waits for the second regroup.