[OM-76932]: Show clear message when webhook url used for webhook is not valid
The error message when the url for webhook is invalid should be clear

The ``URI.create`` method throws an ``IllegalArgumentException`` when an invalid url argument is sent to it. We don't capture this exception properly in webhook probe. Therefore, the message that user sees in the API when calling this method is not clear. The changes here updated the webhook probe code to capture this exception properly and show a proper error message in the API response.