[OM-76693] Updated Prometheus->DataCloud gateway version.
This was done to address a number of problems including entering
`CrashLoopBackOff` when there was no internet connection at the time of
container creation and logging too much when internet connection was lost
after container creation.

Also related to DC-1357 and DC-1358.

Deployed modified version of t8c-operator in a lab appliance and
verified that:

1. When there is an internet connection, DataCloud receives data
   for Graph.
2. When the internet connection is disabled before creating the
   pod, the `datacloud-gateway-graph` container doesn’t crash
   and prints messages following an exponential back-off.
3. When the internet connection is disabled after creating the
   pod, the `datacloud-gateway-graph` container doesn’t crash
   and prints a different message following an exponential back-
   off.

Some messages max-out at a back-off of 10 minutes vs others max-
out at the configured 2 hours, but I don’t think we should delay
shipping the fix because of that.