[OM-76693] Changed datacloud-gateway-graph's configuration.
This was done to lower resource consumption for the container to levels
comparable with those of the tag gateway by disabling the self-monitoring
feature of the graph gateway.

When testing the change in a lab XL appliance, memory consumption
reduced from 350-385MB to 8.5-12.3MB, while CPU consumption from ~1% to
0.02-0.04%.

This is related to DC-1359.