[OM-77585]: Fix an issue with sending action progress update to UI observed in some customers
Make it not required to the API component to make a query to action orchestrator to get the action action ID when it recieves an update on the action

When an action progress happens, a message describing the action update is sent on a kafka topic. Different components listen to this topic and act based on the events. One of these components is API. The API sends these updates to UI on websocket so that the action progress update can be reflected in the UI. The message about the update only has the action instance ID. However, we currently only expose the action stable id the user in the API. Therefore, the API needs to make a call to Action orchestrator to get the stable ID for that action instance id. However, if the action stay in the kafka topic for a long time, the action can be discarded in AO. Therefore, we cannot lookup the action stable ID and that is what the bug is about.

To fix this issue, we will include the action stable ID in the action update notifications. This way there is no need to call AO to get the action stable ID. To do so, we add stable ID to the list of data we keep for Action in tp when send actions to probe and use it when creating update message to be sent in the kafka.

The initial fix for this action extracted the stable ID of the action from the context data of first action. However, there was a issue in the code that added stable ID to context data that resulted in the stable ID not being present for some action types. I fixed that issue. However, sending stable ID as part of context data is not a good idea as it is prone to these issues. Therefore, I added a dedicated field to ActionExecutionRequest to have the action stable ID. I also updated the actionscript probe that uses the stable to look at this field rather than trying to find the stable ID in the context data.