[OM-81065]: Support discovered move automation policy from probe
Enable probes to create an automation policy that manages automation mode of move actions

For vms associated to specific types of Flexera licenses, we want to create policies that prevent those vms to move from their current host. For those vms, we want to disable move actions from probe side. This changes allows that to happen.