[OM-89423] Fixed a bug where Prometheus server wouldn't start in 1.14.3.
When we removed the containers and secret used to send telemetry data to
DataCloud, we forgot to remove the extra volume from the prometheus-
server pod. Since no container is trying to mount the non-existent
volume this works fine in most versions of Kubernetes but produces an
error in 1.14.3 used by some of our oldest XL customers. (e.g. 8.0.0)

Also did a search for mentions of DataCloud in case we've missed
anything else and updated some comments that were found.