[OM-89570] Changed default RBAC for Prometheus and Alert Manager.
By default, Prometheus and Alert Manager will now use namespaced roles
instead of cluster roles. I still need to make cluster roles the default
again for OVAs and avoid hard-coding the namespace name.

For Alert Manager, the required functionality was already there in the
Helm chart. For Prometheus server I had to create a new template for the
role and modify the ones for the other resources.

The new namespaced role for Prometheus was based on the cluster role but
with access to cluster-wide resources removed. We don't need such access
for telemetry.

Testing was done on a Rancher Desktop Kubernetes cluster by using the
namespaced role for t8c-operator and making sure all required resources
are created without an issue and that telemetry is collected without any
errors appearing in Prometheus server logs.