[OM-89570] Modified some scrape jobs to be constrained to namespace.
The built-in telemetry job and some Alert Manager configuration were
modified to look only in the namespace where the XL instance is deployed
in. This was done, so they can work without errors when a namespaced
role is used.

Also modified telemetry scrape job to add a `port` label and to scrape
all endpoints of all services in the namespace.

Finally, Prometheus was enabled by default.