[OM-89570] Moved out-of-the-box scrape jobs behind a feature flag.
These Kubernetes- and XL-related scrape jobs aren't needed in all
deployments and should be off by default in t8c-operator. They are used
in OVA deployments and should be on by default there.

Since Helm's handling of lists is problematic the only practical way to
add and remove jobs easily is to move them to the template (from values)
and put them behind a flag mentioned in values.yaml.