[OM-89570] Simplified RBAC object creation logic for Prometheus chart.
Removed logic to create roles and role bindings for Prometheus server
and Alert Manager in multiple namespaces as well as the ability to use a
pre-defined cluster role with Prometheus server.

Now Prometheus and Alert Manager can either use a cluster role or a
namespaced role in their own namespace. In any case, the role is created
by t8c-operator.

This was done because we don't have the capacity to test all
combinations, and we've never had to use any of the other use cases so
far. In fact, when t8c-operator isn't given a cluster role, it doesn't
have permission to create roles in other namespaces anyway.

I believe the reason these features existed in the upstream chart for
Prometheus is that it's intended to be used directly by the cluster
administrator to deploy Prometheus for monitoring and not by an operator
to deploy Prometheus as a component of a larger system.

Also updated some documentation comments.